SQL and clean-up code
This program allows the user to remove anomalous data from their output .json and .csv files, and further process it by creating both individual SQL libraries for each set of results, as well as a combined one for the whole experiment. At the end of the notebook, the user can also find an example of a query, that can be done with the newly created library.