Data correction and SQL
This code allows the user to clean the experimental json and csv files. The code then allows the user to compile the cleaned data into individual and merged SQL databases. At the end of the notebook, the user can find an example query to search for specific conditions, and test whether the database was created successfully.